+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/api/index.ts
@@ -3,36 +3,6 @@
 
 export {
   createHealthDataAIServices,
-  type HealthDataAIServicesContext,
-  type HealthDataAIServicesClientOptionalParams,
+  HealthDataAIServicesContext,
+  HealthDataAIServicesClientOptionalParams,
 } from "./healthDataAIServicesContext.js";
-export {
-  type OperationsListOptionalParams,
-  type DeidServicesGetOptionalParams,
-  type DeidServicesListByResourceGroupOptionalParams,
-  type DeidServicesListBySubscriptionOptionalParams,
-  type DeidServicesCreateOptionalParams,
-  type DeidServicesUpdateOptionalParams,
-  type DeidServicesDeleteOptionalParams,
-  type PrivateEndpointConnectionsGetOptionalParams,
-  type PrivateEndpointConnectionsCreateOptionalParams,
-  type PrivateEndpointConnectionsDeleteOptionalParams,
-  type PrivateEndpointConnectionsListByDeidServiceOptionalParams,
-  type PrivateLinksListByDeidServiceOptionalParams,
-} from "./options.js";
-export {
-  deidServicesGet,
-  deidServicesListByResourceGroup,
-  deidServicesListBySubscription,
-  deidServicesCreate,
-  deidServicesUpdate,
-  deidServicesDelete,
-} from "./deidServices/index.js";
-export { operationsList } from "./operations/index.js";
-export {
-  privateEndpointConnectionsGet,
-  privateEndpointConnectionsCreate,
-  privateEndpointConnectionsDelete,
-  privateEndpointConnectionsListByDeidService,
-} from "./privateEndpointConnections/index.js";
-export { privateLinksListByDeidService } from "./privateLinks/index.js";