+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { AzureClouds, AzureSupportedClouds } from "./static-helpers/cloudSettingHelpers.js";
 import {
   PageSettings,
   ContinuablePage,
@@ -8,64 +9,72 @@ import {
 } from "./static-helpers/pagingHelpers.js";
 
 export { HealthDataAIServicesClient } from "./healthDataAIServicesClient.js";
-export { restorePoller, type RestorePollerOptions } from "./restorePollerHelpers.js";
+export { restorePoller, RestorePollerOptions } from "./restorePollerHelpers.js";
 export {
-  type PrivateLinkResource,
-  type PrivateLinkResourceProperties,
-  type ProxyResource,
-  type Resource,
-  type SystemData,
-  KnownCreatedByType,
-  type CreatedByType,
-  type PrivateEndpointConnectionResource,
-  type PrivateEndpointConnectionProperties,
-  type PrivateEndpoint,
-  type PrivateLinkServiceConnectionState,
-  KnownPrivateEndpointServiceConnectionStatus,
-  type PrivateEndpointServiceConnectionStatus,
-  KnownPrivateEndpointConnectionProvisioningState,
-  type PrivateEndpointConnectionProvisioningState,
-  type DeidService,
-  type DeidServiceProperties,
-  KnownProvisioningState,
-  type ProvisioningState,
-  type PrivateEndpointConnection,
-  type PublicNetworkAccess,
-  type ManagedServiceIdentity,
-  KnownManagedServiceIdentityType,
-  type ManagedServiceIdentityType,
-  type UserAssignedIdentity,
-  type TrackedResource,
-  type DeidUpdate,
-  type ManagedServiceIdentityUpdate,
-  type DeidPropertiesUpdate,
-  type Operation,
-  type OperationDisplay,
+  Operation,
+  OperationDisplay,
   KnownOrigin,
-  type Origin,
+  Origin,
   KnownActionType,
-  type ActionType,
+  ActionType,
+  ErrorResponse,
+  ErrorDetail,
+  ErrorAdditionalInfo,
+  DeidService,
+  DeidServiceProperties,
+  KnownProvisioningState,
+  ProvisioningState,
+  PrivateEndpointConnection,
+  PrivateEndpointConnectionProperties,
+  PrivateEndpoint,
+  PrivateLinkServiceConnectionState,
+  KnownPrivateEndpointServiceConnectionStatus,
+  PrivateEndpointServiceConnectionStatus,
+  KnownPrivateEndpointConnectionProvisioningState,
+  PrivateEndpointConnectionProvisioningState,
+  PublicNetworkAccess,
+  ManagedServiceIdentity,
+  KnownManagedServiceIdentityType,
+  ManagedServiceIdentityType,
+  UserAssignedIdentity,
+  Sku,
+  SkuTier,
+  Resource,
+  SystemData,
+  KnownCreatedByType,
+  CreatedByType,
+  TrackedResource,
+  DeidUpdate,
+  ManagedServiceIdentityUpdate,
+  DeidPropertiesUpdate,
+  PrivateEndpointConnectionResource,
+  ProxyResource,
+  PrivateLinkResource,
+  PrivateLinkResourceProperties,
   KnownVersions,
 } from "./models/index.js";
+export { HealthDataAIServicesClientOptionalParams } from "./api/index.js";
 export {
-  type HealthDataAIServicesClientOptionalParams,
-  type OperationsListOptionalParams,
-  type DeidServicesGetOptionalParams,
-  type DeidServicesListByResourceGroupOptionalParams,
-  type DeidServicesListBySubscriptionOptionalParams,
-  type DeidServicesCreateOptionalParams,
-  type DeidServicesUpdateOptionalParams,
-  type DeidServicesDeleteOptionalParams,
-  type PrivateEndpointConnectionsGetOptionalParams,
-  type PrivateEndpointConnectionsCreateOptionalParams,
-  type PrivateEndpointConnectionsDeleteOptionalParams,
-  type PrivateEndpointConnectionsListByDeidServiceOptionalParams,
-  type PrivateLinksListByDeidServiceOptionalParams,
-} from "./api/index.js";
+  DeidServicesDeleteOptionalParams,
+  DeidServicesUpdateOptionalParams,
+  DeidServicesCreateOptionalParams,
+  DeidServicesListBySubscriptionOptionalParams,
+  DeidServicesListByResourceGroupOptionalParams,
+  DeidServicesGetOptionalParams,
+} from "./api/deidServices/index.js";
+export { OperationsListOptionalParams } from "./api/operations/index.js";
 export {
-  type DeidServicesOperations,
-  type OperationsOperations,
-  type PrivateEndpointConnectionsOperations,
-  type PrivateLinksOperations,
+  PrivateEndpointConnectionsListByDeidServiceOptionalParams,
+  PrivateEndpointConnectionsDeleteOptionalParams,
+  PrivateEndpointConnectionsCreateOptionalParams,
+  PrivateEndpointConnectionsGetOptionalParams,
+} from "./api/privateEndpointConnections/index.js";
+export { PrivateLinksListByDeidServiceOptionalParams } from "./api/privateLinks/index.js";
+export {
+  DeidServicesOperations,
+  OperationsOperations,
+  PrivateEndpointConnectionsOperations,
+  PrivateLinksOperations,
 } from "./classic/index.js";
-export { type PageSettings, type ContinuablePage, type PagedAsyncIterableIterator };
+export { PageSettings, ContinuablePage, PagedAsyncIterableIterator };
+export { AzureClouds, AzureSupportedClouds };