+++ sdk/healthdataaiservices/arm-healthdataaiservices/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Release History
+
+## 1.1.0 (Unreleased)
+
+### Features Added
+
+  - Added SKU support to `DeidService` resource model using ARM common-types
+  - Added optional `sku` property to `DeidUpdate` patch model
     
 ## 1.0.0 (2024-11-25)